+++ resource_cache/resources/resource.cpp
@@ -1,10 +1,12 @@
 #include "resource.hpp"
 
-Resource::Resource(std::string name, uint8_t* buffer, unsigned int size,
-		std::shared_ptr<IResourceStore> store,
-		std::shared_ptr<ResourceCache> cache)
-	: _name(std::move(name)), _buffer(buffer), _size(size),
-	_store(std::move(store)), _cache(std::move(cache))
+Resource::Resource(std::string name, uint8_t* buffer, uint64_t size,
+	std::shared_ptr<IResourceStore> store, std::shared_ptr<ResourceCache> cache)
+	: _name(std::move(name)),
+	  _buffer(buffer),
+	  _size(size),
+	  _store(std::move(store)),
+	  _cache(std::move(cache))
 {}
 
 Resource::~Resource()
